backend/api/urls.py
from django.urls import include, path
from rest_framework.routers import DefaultRouter
from users.views import CustomUserViewSet
from api.views import (IngredientsViewSet, TagsViewSet, RecipesViewSet, IngredientsAmountViewSet)

router = DefaultRouter()

router.register(r'users', CustomUserViewSet, basename='users')
router.register(r'tags', TagsViewSet, basename='tags')
router.register(r'ingredients', IngredientsViewSet, basename='ingredients')
router.register(r'recipes', RecipesViewSet, basename='recipes')
router.register(r'ingredientsamount', IngredientsAmountViewSet, basename='ingredientsamount')


urlpatterns = [
    path('', include(router.urls)),
    path('auth/', include('djoser.urls')),
    path('auth/', include('djoser.urls.authtoken')),
]