backend/requirements.txt
gunicorn==20.1.0
Django
djangorestframework==3.15.2
djoser==2.2.3
Pillow==10.4.0
django-filter==23.1